python-fastapi/pyproject.toml
[tool.poetry]
name = "python-app"
version = "0.1.0"
description = "A FastAPI app with Kafka and Postgres"
authors = ["Mouad Kondah <mouad.kondah@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.70.0"
uvicorn = "==0.30.3"
kafka-python = "^2.0.2"
psycopg2-binary = "^2.9.1"
aiokafka = "^0.11.0"
asyncpg = "^0.29.0"
opentelemetry-api = "^1.26.0"
opentelemetry-instrumentation = "^0.47b0"
opentelemetry-exporter-otlp = "^1.26.0"
opentelemetry-instrumentation-fastapi = "^0.47b0"
opentelemetry-sdk = "^1.26.0"
opentelemetry-instrumentation-asyncpg = "^0.47b0"
opentelemetry-instrumentation-kafka-python = "^0.47b0"
httpx = "^0.27.0"
opentelemetry-instrumentation-httpx = "^0.47b0"
pyyaml = "^6.0.2"

